+++ src/journal/journald-wall.c
@@ -6,7 +6,7 @@
 #include "journald-wall.h"
 #include "process-util.h"
 #include "string-util.h"
-#include "utmp-wtmp.h"
+#include "wall.h"
 
 void server_forward_wall(
                 Server *s,
@@ -48,7 +48,7 @@ void server_forward_wall(
         } else
                 l = message;
 
-        r = utmp_wall(l, "systemd-journald", NULL, NULL, NULL);
+        r = wall(l, "systemd-journald", NULL, NULL, NULL);
         if (r < 0)
                 log_debug_errno(r, "Failed to send wall message: %m");
 }
