+++ src/login/meson.build
@@ -26,7 +26,7 @@ liblogind_core_sources = files(
         'logind-session.c',
         'logind-user-dbus.c',
         'logind-user.c',
-        'logind-utmp.c',
+        'logind-wall.c',
 )
 
 liblogind_core_sources += [logind_gperf_c]
