+++ src/shared/utmp-wtmp.h
@@ -18,13 +18,6 @@ int utmp_put_runlevel(int runlevel, int previous);
 int utmp_put_dead_process(const char *id, pid_t pid, int code, int status);
 int utmp_put_init_process(const char *id, pid_t pid, pid_t sid, const char *line, int ut_type, const char *user);
 
-int utmp_wall(
-        const char *message,
-        const char *username,
-        const char *origin_tty,
-        bool (*match_tty)(const char *tty, bool is_local, void *userdata),
-        void *userdata);
-
 static inline bool utxent_start(void) {
         setutxent();
         return true;
@@ -52,14 +45,6 @@ static inline int utmp_put_dead_process(const char *id, pid_t pid, int code, int
         return 0;
 }
 static inline int utmp_put_init_process(const char *id, pid_t pid, pid_t sid, const char *line, int ut_type, const char *user) {
-        return 0;
-}
-static inline int utmp_wall(
-                const char *message,
-                const char *username,
-                const char *origin_tty,
-                bool (*match_tty)(const char *tty, bool is_local, void *userdata),
-                void *userdata) {
         return 0;
 }
